Finalize change
