+++ src/app/component/config-model-list/index.js
@@ -1,0 +1,1 @@
+export {default} from './config-model-list'
