--- src/app/component/model-upload-list/_model-upload-list.scss
@@ -1,13 +1,0 @@
-.ModelUploadList {
-  list-style: none;
-  padding: 0;
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-gap: $space-xl;
-  margin: 0;
-
-  &__item {
-    height: $upload-model-item-height;
-    min-width: 0;
-  }
-}
