--- src/app/component/model-upload-list/index.js
@@ -1,1 +1,0 @@
-export {default} from './model-upload-list'
