+++ src/app/component/upload-model-list/_upload-model-list.scss
@@ -1,68 +1,13 @@
 .UploadModelList {
-  position: relative;
+  list-style: none;
+  padding: 0;
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-gap: $space-xl;
+  margin: 0;
 
   &__item {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    transition: transform 0.3s 0.3s ease-in-out, opacity 0.3s ease-in-out;
-
-    // First item to exit
-    &.PortalTransition-appear-active {
-      animation: UploadModelList-phase1-exit-first 0.3s cubic-bezier(0.25, 0.1, 0.25, 1),
-        UploadModelList-phase2-exit-first 0.3s 0.3s cubic-bezier(0.9, -0.02, 0.58, 1) forwards !important;
-    }
-
-    // item to exit expect the first
-    &.PortalTransition-appear-active + &.PortalTransition-appear-active {
-      animation: UploadModelList-phase1-exit 0.3s cubic-bezier(0.25, 0.1, 0.25, 1),
-        UploadModelList-phase2-exit 0.3s 0.3s cubic-bezier(0.9, -0.02, 0.58, 1) forwards !important;
-    }
-
-    &.PortalTransition-appear-done {
-      display: none;
-    }
-
-    &.ItemTransition-exit-active,
-    &.ItemTransition-exit {
-      opacity: 0;
-    }
-  }
-
-  @keyframes UploadModelList-phase1-exit-first {
-    to {
-      transform: translateY(0);
-      opacity: 1;
-    }
-  }
-
-  @keyframes UploadModelList-phase1-exit {
-    to {
-      transform: translateY($space-base) scale(0.95);
-      opacity: 1;
-    }
-  }
-
-  @keyframes UploadModelList-phase2-exit-first {
-    from {
-      transform: translateY(0);
-      opacity: 1;
-    }
-    to {
-      transform: translate(60vw, -200px);
-      opacity: 0;
-    }
-  }
-
-  @keyframes UploadModelList-phase2-exit {
-    from {
-      transform: translateY($space-base) scale(0.95);
-      opacity: 1;
-    }
-    to {
-      transform: translate(60vw, -200px) scale(0.95);
-      opacity: 0;
-    }
+    height: $upload-model-item-height;
+    min-width: 0;
   }
 }
